Add files via upload
Near-final version of the program, utilizing categorize plug in instead of single stimulus plug in. The program now cycles through a pseudorandomized list of stimuli and provides feedback in the form of a line of text. More pseudo-randomized timelines need to be added, as well as the correct reward amounts, and a prompt at the opening of the experiment to collect subject information